+++ src/domain/usecases/add-account.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from './models/account'
+import { AccountModel } from '../models/account'
 
 export interface AddAccountModel {
   name: string
